+++ docs/index.html
@@ -1486,7 +1486,7 @@
       <p id="vocab-last-generated" class="note">
       This vocabulary was last generated
       <span id="vocab-last-generated-date">
-      Thursday, October 15, 2020 8:46 AM
+      Thursday, October 15, 2020 8:47 AM
       </span>
             </p>
     </section>
